requirements.txt
fastapi
uvicorn
numpy
scikit-learn
python-multipart
python-dotenv
